Fix `HEAP_INIT_SLOTS` handling
The meaning of `INIT_SLOTS` is to
start a process with at least that number
of free slots to reduce GC cycles.

However, up to now the `INIT_SLOTS`
would limit how many free slots would
be able to be freed.

This is dues the `INIT_SLOTS` affecting
the `MAX_FREE_SLOTS` present.